refactor: lazy load members
