+++ packages/features/ee/teams/components/DeleteBulkTeamMembers.tsx
@@ -18,7 +18,6 @@ export default function DeleteBulkTeamMembers({ users, onRemove, isOrg, teamId }
   const deleteMutation = trpc.viewer.teams.removeMember.useMutation({
     async onSuccess() {
       await utils.viewer.teams.get.invalidate();
-      await utils.viewer.teams.lazyLoadMembers.invalidate();
       await utils.viewer.eventTypes.invalidate();
       await utils.viewer.organizations.listMembers.invalidate();
       await utils.viewer.organizations.getMembers.invalidate();
