+++ packages/features/ee/teams/components/MemberInvitationModal.tsx
@@ -513,7 +513,7 @@ export const MemberInvitationModalWithoutMembers = ({
           {
             onSuccess: async (data) => {
               await utils.viewer.teams.get.invalidate();
-              await utils.viewer.teams.lazyLoadMembers.invalidate();
+              await utils.viewer.teams.listMembers.invalidate();
               await utils.viewer.organizations.getMembers.invalidate();
               hideInvitationModal();
