+++ packages/features/settings/outOfOffice/CreateOrEditOutOfOfficeModal.tsx
@@ -41,6 +41,7 @@ export const CreateOrEditOutOfOfficeEntryModal = ({
   const { t } = useLocale();
   const utils = trpc.useUtils();
 
+  // TODO: support infinite scrolling
   const { data: listMembers } = trpc.viewer.teams.listMembers.useQuery({});
   const me = useMeQuery();
   const memberListOptions: {
