+++ packages/trpc/server/routers/viewer/teams/listMembers.schema.ts
@@ -1,7 +1,10 @@
 import { z } from "zod";
 
 export const ZListMembersInputSchema = z.object({
-  teamIds: z.number().array().optional(),
+  teamId: z.number(),
+  limit: z.number().default(10),
+  searchTerm: z.string().optional(),
+  cursor: z.number().optional().nullable(),
 });
 
 export type TListMembersInputSchema = z.infer<typeof ZListMembersInputSchema>;
